refactor(stdio): move byte framing into a shared module
`FrameReader`, `InputFrame` and `decode_input_frame` were private to the
server's stdio transport. The client transport reads the same wire format
and needs the same byte-level framing, so they move to a crate-private
`framing` module rather than being written twice.

No behaviour change. The pieces that are specific to the server's answer,
`UNDECODABLE_FRAME` and `undecodable_frame_response`, stay behind, because
the `-32700` is the part the two ends do not share.

The frame-reading tests move with the code. One is added there: that a
partial frame survives a cancelled read. Both ends poll `next_frame` inside
a `select!`, so that property belongs with the reader rather than with
either caller.
